Adjust docs
